text-bi-llm-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
httpx
SQLAlchemy
pymysql
